[REFACTOR] #1 패키지 구조 분리 및 토스트 확장 함수
